Refactored tests und unnötige Checks auf Fehlermeldungsinhalt entfernt
