+++ src/test/java/de/prog3/projektarbeit/data/factory/PlayerFactoryTests.java
@@ -49,62 +49,48 @@ public class PlayerFactoryTests {
     @DisplayName("Build with empty first name throws ValidationException")
     void build_emptyFirstName_throwsException() {
         playerFactory.setFirstName(""); // Ungültiger Vorname
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Fehlender Vorname"),
-                "Die Fehlermeldung sollte 'Fehlender Vorname' enthalten.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 
     @Test
     @DisplayName("Build with blank last name throws ValidationException")
     void build_blankLastName_throwsException() {
         playerFactory.setLastName("   "); // Nur Leerzeichen als Nachname
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Fehlender Nachname"),
-                "Die Fehlermeldung sollte 'Fehlender Nachname' enthalten.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 
     @Test
     @DisplayName("Build with null date of birth throws ValidationException")
     void build_nullDateOfBirth_throwsException() {
         playerFactory.setDateOfBirth(null);
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Fehlender Geburtstag"),
-                "Die Fehlermeldung sollte 'Fehlender Geburtstag' enthalten.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 
     @Test
     @DisplayName("Build with invalid number (0) throws ValidationException")
     void build_invalidNumberZero_throwsException() {
         playerFactory.setNumber(0); // Ungültige Nummer: 0
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Spielernummer muss größer als 0 und kleiner als 100"),
-                "Die Fehlermeldung sollte auf die Spielernummer hinweisen.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 
     @Test
     @DisplayName("Build with invalid number (>= 100) throws ValidationException")
     void build_invalidNumberTooHigh_throwsException() {
         playerFactory.setNumber(100); // Ungültige Nummer: 100
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Spielernummer muss größer als 0 und kleiner als 100"),
-                "Die Fehlermeldung sollte auf die Spielernummer hinweisen.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 
     @Test
     @DisplayName("Build with null positions throws ValidationException")
     void build_nullPositions_throwsException() {
         playerFactory.setPositions(null);
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Spieler muss mindestens eine Position haben"),
-                "Die Fehlermeldung sollte 'Spieler muss mindestens eine Position haben' enthalten.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 
     @Test
     @DisplayName("Build with empty positions list throws ValidationException")
     void build_emptyPositions_throwsException() {
         playerFactory.setPositions(new ArrayList<>());
-        ValidationException ex = assertThrows(ValidationException.class, () -> playerFactory.build());
-        assertTrue(ex.getMessage().contains("Spieler muss mindestens eine Position haben"),
-                "Die Fehlermeldung sollte 'Spieler muss mindestens eine Position haben' enthalten.");
+        assertThrows(ValidationException.class, () -> playerFactory.build());
     }
 }
