+++ src/test/java/de/prog3/projektarbeit/events/TournamentManagerTests.java
@@ -2,7 +2,7 @@ package de.prog3.projektarbeit.events;
 
 import de.prog3.projektarbeit.data.objects.Team;
 import de.prog3.projektarbeit.eventHandling.events.data.Match;
-import de.prog3.projektarbeit.eventHandling.events.data.team.TournamentManager;
+import de.prog3.projektarbeit.data.TournamentManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -22,7 +22,7 @@ public class TournamentManagerTests {
     public void testCreateTournament() {
         tournamentManager.createTournament("Summer Cup");
         assertEquals(1, tournamentManager.getTournaments().size(), "There should be one tournament created" );
-        assertEquals("Summer Cup", tournamentManager.getTournaments().get(0).getName(), "Tournament name should be 'Summer Cup'");
+        assertEquals("Summer Cup", tournamentManager.getTournaments().getFirst().getName(), "Tournament name should be 'Summer Cup'");
     }
 
     @Test
@@ -30,8 +30,8 @@ public class TournamentManagerTests {
         tournamentManager.createTournament("Summer Cup");
         Team team = new Team("Team A");
         tournamentManager.addTeamToTournament("Summer Cup", team);
-        assertEquals(1, tournamentManager.getTournaments().get(0).getTeams().size(), "There should be one team in the tournament");
-        assertEquals("Team A", tournamentManager.getTournaments().get(0).getTeams().get(0).getName(), "Team name should be 'Team A'");
+        assertEquals(1, tournamentManager.getTournaments().getFirst().getTeams().size(), "There should be one team in the tournament");
+        assertEquals("Team A", tournamentManager.getTournaments().getFirst().getTeams().getFirst().getName(), "Team name should be 'Team A'");
     }
 
     @Test
@@ -41,9 +41,9 @@ public class TournamentManagerTests {
         Team team2 = new Team("Team B");
         Match match = new Match(team1, team2, "2025-06-15");
         tournamentManager.addMatchToTournament("Summer Cup", match);
-        assertEquals(1, tournamentManager.getTournaments().get(0).getMatches().size(), "There should be one match in the tournament");
-        assertEquals(team1, tournamentManager.getTournaments().get(0).getMatches().get(0).getHomeTeam(), "Home team should be 'Team A'");
-        assertEquals(team2, tournamentManager.getTournaments().get(0).getMatches().get(0).getAwayTeam(), "Away team should be 'Team B'");
+        assertEquals(1, tournamentManager.getTournaments().getFirst().getMatches().size(), "There should be one match in the tournament");
+        assertEquals(team1, tournamentManager.getTournaments().getFirst().getMatches().getFirst().getHomeTeam(), "Home team should be 'Team A'");
+        assertEquals(team2, tournamentManager.getTournaments().getFirst().getMatches().getFirst().getAwayTeam(), "Away team should be 'Team B'");
     }
 
     @Test
@@ -51,6 +51,6 @@ public class TournamentManagerTests {
         Exception exception = assertThrows(IllegalArgumentException.class, () -> {
             tournamentManager.addTeamToTournament("Nonexistent Cup", new Team("Team A"));
         });
-        assertEquals("Tournament not found: Nonexistent Cup", exception.getMessage());
+        assertEquals("Tournament not found Nonexistent Cup does not exist", exception.getMessage());
     }
 }
